Move TryIt into own package, aka monorepo hell
