+++ jest.config.js
@@ -30,7 +30,8 @@ module.exports = {
   moduleDirectories: ['./node_modules', './packages'],
   moduleFileExtensions: ['js', 'jsx', 'ts', 'tsx', 'json', 'node'],
   moduleNameMapper: {
-    '@looker/(.+)$': '<rootDir>/packages/$1/src',
+    '@looker/sdk/lib/browser$': '<rootDir>/../sdk',
+    '@looker/((?!components|design|icons)(.+))$': '<rootDir>/packages/$1/src',
     '\\.(jpg|jpeg|png|gif|eot|otf|webp|svg|ttf|woff|woff2|mp4|webm|wav|mp3|m4a|aac|oga)$':
       '<rootDir>/config/jest/fileMock.js',
   },
