+++ packages/api-explorer/package.json
@@ -14,6 +14,7 @@
     "url": "https://github.com/looker-open-source/sdk-codegen/issues"
   },
   "scripts": {
+    "test": "jest -- src/",
     "develop": "webpack-dev-server --https",
     "docs": "typedoc --mode file --out docs"
   },
@@ -75,6 +76,7 @@
     "@looker/sdk": "^0.3.3-alpha.0",
     "@looker/sdk-codegen": "^0.1.2-alpha.18+551e4fe",
     "@looker/sdk-codegen-utils": "^0.1.0-alpha.2",
+    "@looker/try-it": "0.1.0-alpha",
     "@testing-library/jest-dom": "^5.5.0",
     "@types/lodash": "^4.14.144",
     "@types/node": "^13.13.4",
